+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Each tagged release MUST have a section 2 heading starting at the time of releas
   - Added the ability to run multiple seperate development environments via the `--context` flag (T301002)
   - Added the `--force-recreate` command to service `create` commands (T313411)
   - Added `docker mediawiki get-code` command for fetching MediaWiki, skins and extensions from Gerrit
+  - Added notes about Windows hosts files when using WSL and the `docker hosts` commands
   - Removed code clone wizard from `docker mediawiki` command startup, insteading prompting users to use `docker mediawiki get-code`
   - Image updates:
     - postgres `postgres:13.6` -> `postgres:13.9`